requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
distro==1.5.0
idna==2.9
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
pyocr==0.7.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
tabula-py==2.1.0
urllib3==1.25.9
